ci: replace Droid review with deterministic gate
