+++ scripts/tests/test_land_pr.py
@@ -61,119 +61,10 @@ class LandPRTests(unittest.TestCase):
         self.assertTrue(ok)
         self.assertEqual(reason, "")
 
-    def test_check_droid_finished_accepts_finished_review(self):
-        ok, reason = land_pr.check_droid_finished(
-            [
-                {
-                    "user": {"login": "factory-droid[bot]", "type": "Bot"},
-                    "body": "**Droid finished @alice's task**\n\nNo actionable findings.",
-                    "url": "https://comment",
-                }
-            ]
-        )
-        self.assertTrue(ok)
-        self.assertEqual(reason, "")
-
-    def test_check_droid_finished_accepts_progress_boilerplate_in_finished_review(self):
-        ok, reason = land_pr.check_droid_finished(
-            [
-                {
-                    "user": {"login": "factory-droid[bot]", "type": "Bot"},
-                    "body": "**Droid finished @alice's task**\n\nDroid is reviewing code and running a security check...\n\nReview approved.",
-                    "url": "https://comment",
-                }
-            ]
-        )
-        self.assertTrue(ok)
-        self.assertEqual(reason, "")
-
-    def test_check_droid_finished_accepts_review_complete_comment(self):
-        ok, reason = land_pr.check_droid_finished(
-            [
-                {
-                    "user": {"login": "factory-droid[bot]", "type": "Bot"},
-                    "body": "Review complete for PR #42. No actionable findings.",
-                    "url": "https://comment",
-                }
-            ]
-        )
-        self.assertTrue(ok)
-        self.assertEqual(reason, "")
-
-    def test_check_droid_finished_rejects_bare_login_spoof(self):
-        ok, reason = land_pr.check_droid_finished(
-            [
-                {
-                    "user": {"login": "factory-droid", "type": "User"},
-                    "body": "**Droid finished @alice's task**\n\nLGTM.",
-                    "url": "https://comment",
-                }
-            ]
-        )
-        self.assertFalse(ok)
-        self.assertIn("missing", reason)
-
-    def test_check_droid_finished_rejects_error(self):
-        ok, reason = land_pr.check_droid_finished(
-            [
-                {
-                    "user": {"login": "factory-droid[bot]", "type": "Bot"},
-                    "body": "**Droid encountered an error**\n\nFailed to checkout PR branch.",
-                    "url": "https://comment",
-                }
-            ]
-        )
-        self.assertFalse(ok)
-        self.assertIn("error", reason)
-
-    def test_check_droid_finished_rejects_in_progress(self):
-        ok, reason = land_pr.check_droid_finished(
-            [
-                {
-                    "user": {"login": "factory-droid[bot]", "type": "Bot"},
-                    "body": "Droid is reviewing code and running a security check...",
-                    "url": "https://comment",
-                }
-            ]
-        )
-        self.assertFalse(ok)
-        self.assertIn("in progress", reason)
-
-    def test_check_droid_finished_accepts_superseded_error_then_finished_review(self):
-        ok, reason = land_pr.check_droid_finished(
-            [
-                {
-                    "user": {"login": "factory-droid[bot]", "type": "Bot"},
-                    "body": "**Superseded Droid error** -- rerun completed successfully.",
-                    "url": "https://old-comment",
-                },
-                {
-                    "user": {"login": "factory-droid[bot]", "type": "Bot"},
-                    "body": "**Droid finished @alice's task**\n\nLGTM.",
-                    "url": "https://new-comment",
-                },
-            ]
-        )
-        self.assertTrue(ok)
-        self.assertEqual(reason, "")
-
-    def test_check_droid_finished_rejects_stale_finished_review(self):
-        ok, reason = land_pr.check_droid_finished(
-            [
-                {
-                    "user": {"login": "factory-droid[bot]", "type": "Bot"},
-                    "body": "**Droid finished @alice's task**\n\nLGTM.",
-                    "url": "https://old-comment",
-                },
-                {
-                    "user": {"login": "factory-droid[bot]", "type": "Bot"},
-                    "body": "Droid posted a non-terminal status update.",
-                    "url": "https://new-comment",
-                },
-            ]
-        )
-        self.assertFalse(ok)
-        self.assertIn("not a finished review", reason)
+    def test_default_required_checks_use_deterministic_review(self):
+        self.assertIn("deterministic-review", land_pr.DEFAULT_REQUIRED_CHECKS)
+        self.assertNotIn("droid-review", land_pr.DEFAULT_REQUIRED_CHECKS)
+        self.assertNotIn("droid-review-preflight", land_pr.DEFAULT_REQUIRED_CHECKS)
 
     def test_check_no_active_review_threads_accepts_empty_threads(self):
         ok, reason = land_pr.check_no_active_review_threads([])
